Remove Game-Drawer cyclic dependency, rename Grid as Maze, move score draw and alerts into Drawer and move Drawer.mode into MainLoop.mode
Now drawer has all the drawer tools used by the MainLoop, main loop contains the global state of the game and Game is responsible of bootstrapping the whole app, i.e. Game create Drawer and de MainLoop, configure the drawer with the environment specific widgets, and set this to the MainLoop. The rest of the game parts are started by the MainLoop.